Enable external interrupts for stm32g030, stm32g031, and stm32g041
For non stm32g07x and stm32g081 chips, the exti wakeup() method is
hidden behind a conditional compilation attribute. This means an
external interrupt wakeup is never unmasked and the EXTIx_x vector
handler is never called.

Amend the wakeup() and unlisten() methods to work with chips that
only have an imr1 register.

Compile out Event enum members that do not exist on some chips so a
program doesn't try to set a reserved bit in the imr1 register.

Set a constant for the upper signal trigger limit per chip so a
program doesn't try to set a reserved bit in the rtsr1, ftsr1, rpr1,
and fpr1 registers.