UART: Make Config a template param
This enables differentiation between 'Full', 'Basic' and 'Low Power'
feature sets that are available on the STM32G0 parts. They could be
considered as different peripheral, however they share a lot of
funcationality.

This commit splits up the code in a 'shared' part and separate 'full'
and 'basic' implementations that need different config objects and have
specific initialization.