Support PWM output on inverted CHx pins
And add a few missing non-inverted TIM15 pins.

The inverted outputs are the same as the non-inverted outputs, except they are
inverted of course. Since inverted and non-inverted outputs never share the same
physical pin, I think it's reasonable that the inversion should be considered at
the hardware level. Therefore we don't need to build an additional abstraction
for inversion in the HAL, and the straightforward implementation here is just
fine.